+++ tests/_async_compat/__init__.py
@@ -17,7 +17,9 @@
 from functools import wraps as _wraps
 
 from .mark_decorator import (
+    async_fixture,
     AsyncTestDecorators,
+    fixture,
     mark_async_test,
     mark_sync_test,
     TestDecorators,
@@ -27,6 +29,8 @@ from .mark_decorator import (
 __all__ = [
     "AsyncTestDecorators",
     "TestDecorators",
+    "async_fixture",
+    "fixture",
     "mark_async_test",
     "mark_sync_test",
     "wrap_async",